test: move parser type node unit tests to snaptest
